chore: fix rust code to comply with pr checks
