+++ sentinel/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2024"
 
 
 [lints.clippy]
-all = "forbid"
-pedantic = "warn"
-nursery = "warn"
+pedantic = "deny"
+complexity = "deny"
+perf = "deny"
+style = "deny"
